+++ MagicalLifeServerShell/API/Commands/NewGame.cs
@@ -1,6 +1,7 @@
 ﻿using MagicalLifeAPI.Networking;
 using MagicalLifeAPI.Networking.Server;
 using MagicalLifeAPI.Registry;
+using MagicalLifeAPI.Registry.ItemRegistry;
 using MagicalLifeAPI.World.Data;
 using MagicalLifeServer;
 using MagicalLifeServer.ServerWorld.World_Generation.Generators;